Fixed Sheet1 and adjusted sheet3
Sheet1: Fixed wrong column being checked for duplicates. Fixed Msgbox not showing what value it was duplicated.
Sheet 3: Duplicate check was going very slow and making the excel blink in and out. Broke the duplicate check into two different parts which resovled the issue of speed and blinking. Code wise there is a lot of duplicated code.  Need to find a way to streamline the amout of code.